Mesclando informações do cliente com as empresas para melhorar a requisição.
