--- get_empresas/views/clientes.py
@@ -1,23 +1,0 @@
-from django.http import JsonResponse
-from odbc_reader.services import fetch_data
-from rest_framework.views import APIView
-from rest_framework.permissions import IsAuthenticated
-def rename_key(list, mapping):
-    for item in list:
-        for chave_antiga, chave_nova in mapping.items():
-            if chave_antiga in item:
-                item[chave_nova] = item.pop(chave_antiga)
-    return list
-
-class get_clientes(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        if request.method != 'POST':
-            return JsonResponse({"error": "Método não permitido, use POST"}, status=405)
-        
-        query = 'SELECT * FROM bethadba.HRCLIENTE'
-        result = fetch_data(query)
-
-        
-        return JsonResponse({"Empresas": result}, safe=False)
